Basic vise testing
